test added
- test for checkout view added
- test for charity list view added
- test for change activity added
- min and max amount crash fixed

+++ charity/src/main/java/android/omise/charity/domain/model/Donation.kt
@@ -6,5 +6,5 @@ data class Donation(
 
     @SerializedName("name") var donationName: String,
     @SerializedName("token") var donationToken: String,
-    @SerializedName("amount") var donationAmount: Int
+    @SerializedName("amount") var donationAmount: Long
 )
